[Implementation] Added accessToken generation from refreshToken
A new accessToken is created and returned in the response body. The token is created from the encoded username in the refreshToken. Errors are thrown if the refreshToken is invalid or the username doesn't equal the one in the database. Also added getJWTToken method to generate a new JWT token.